scripts/requirements.txt
anthropic
python-dotenv
requests
pydantic

langchain
sentence_transformers